+++ tests/Fixtures/counting-php-command.php
@@ -1,13 +1,15 @@
 <?php
 
-$logFile = $argv[1] ?? null;
+/** @var list<string> $arguments */
+$arguments = $_SERVER['argv'] ?? [];
+$logFile = $arguments[1] ?? null;
 if (!is_string($logFile)) {
     exit(1);
 }
 
 file_put_contents($logFile, "start\n", \FILE_APPEND | \LOCK_EX);
 $process = proc_open(
-    [\PHP_BINARY, ...array_slice($argv, 2)],
+    [\PHP_BINARY, ...array_slice($arguments, 2)],
     [0 => \STDIN, 1 => \STDOUT, 2 => \STDERR],
     $pipes,
 );